edit/play: per-bar tempo map, wait-mode fixes, dense-chart perf, bar ripple tools (#271)
Lands the editor/play work developed on this branch (Aug 17–18), which
had been pushed but never opened as a PR.

- **Per-bar tempo map** (`RecordingMeta.bar_starts`): bar lines can
follow a performance whose tempo breathes. Empty = the old uniform grid,
so every existing bundle is unaffected. Per-bar tools: `nudge_bar_tempo`
(re-times notes), `nudge_bar_length` (grid-only), `insert_bar` /
`remove_bar` (ripple), `nudge_tail`.
- Signed backing offset (`audio_start_us: i64`) with a silent lead-in.
- Wait-mode fixes (fresh-strike requirement, freeze at the note onset),
play honours the piece's hand split.
- Dense-chart performance: viewport culling (`cull.ts`), notes-stripped
`meta` events for note-invariant actions.
- `midi_rescan` host command.

This is the foundation for adaptive-BPM follow-up work (loading a
detected tempo map, and making quantize/metronome/highway follow it).

Gate run locally: fmt, clippy (workspace minus tauri), workspace tests,
`rockcraft-tauri` tests on the Windows host, frontend `tsc --noEmit` +
vitest. Note:
`state::tests::write_bundle_in_place_survives_relative_bundle_dir` (from
#265, already on main) is flaky under parallel test threads because it
calls `set_current_dir`; unrelated to this PR.


---------

+++ crates/core/src/segment.rs
@@ -83,7 +83,7 @@ pub fn slice_segment(
 
     let backing = backing.map(|b| BackingTrack {
         file: b.file.clone(),
-        audio_start_us: b.audio_start_us + seg.start_us,
+        audio_start_us: b.audio_start_us + seg.start_us as i64,
     });
     let video = video.map(|v| BackgroundVideo {
         file: v.file.clone(),
